Add report format specifications to WorkerManager
This allows to specify the report formats to be used at different times
of the WorkerManager's operation. For example, when a task creates a
callback, the format specified in rf_spec['task_finished'] is called.
Each of the rf_spec entries can also be lists, such that multiple report
formats can be invoked without changing anything in the code.