Extend WorkerManager to better handle interrupts.
Now have a grace period to wait for workers to shut down; after that
time, the workers will be killed. There is the option to abort this
grace period and kill the workers directly. Furthermore, it is also
possible to _not_ exit after the KeyboardInterrupt but just continue
with the rest of the code ...