Improve error messages on missing model binary
The Multiverse tests more consistently for a proper executable, directly
in the `_prepare_executable` method, i.e. as early as possible.
Additionally, it also checks for the correct access rights, i.e. that
execution is even possible.

Furthermore, the Task error message is adjusted to be more general.

All corresponding tests were previously missing and are added by this
commit.