Address interpolation issues in CA state plot
By allowing to pass custom parameters to imshow, interpolation can be
defaulted to `none`, thus removing artefacts from the output.

This also sets the default DPI back to 96 in order to speed up plotting,
and sets the `figure.constrained_layout.use` RC parameter to get a
better crop on the figure and the animation.