Let WorkerManager more robustly detect stopped tasks
Stopped meaning that a stop condition triggered a custom signal that led
to them terminating. This should not only work if the signal is expected
and handled in the backend (as now implemented in `utopya_backend`) but
also for other models that do not have such handling. The latter will
leave the exit code in the way they receive it, so the `WorkerManager`
has to take that into account.