Fix & expand imshow_hexagonal and caplot integration
The imshow_hexagonal standalone function is able to handle a large
number of combinations for offset modes, hexagon orientation as well as
boundary modes. All tested and visually confirmed that they work.

Now, the interfaces between imshow and imshow_hexagonal are more
aligned.
This gets rid of a lot of tedious workaround solutions and instead uses
the collection-focussed options to update facecolors. This also makes
the colorbar implementation much more robust and simpler.