Migrate docs from MyST-Markdown back to RST
The MyST syntax is fine and all ... but especially for nested directives
it gets really tedious. On top of that, it feels a lot like a hack,
which shows e.g. when not having proper syntax highlighting in code
editors, which can be very confusing.

Benefit of using reStructuredText is also that it's consistent with the
Utopia and dantro user manuals, making it easier to migrate
documentation pages back and forth.