Fix error in counting of "finished" tasks in WorkerManager
Restructure this into the TaskList-based approach, which is storing a
few references to tasks once they have been worked on or were skipped.

Main problem was: tasks that don't start a worker process will directly
exit and invoke their callback (and their report), so the previously
used `_num_finished_tasks` variable was not correct. The fix was to use
the callback function to increase these counters... This also solved
problems in the computation of the progress bar ticks.