Rework re-running of simulations via DistributedMultiverse
Instead of NoWorkTask, now uses the task-skipping functionality, which
is simpler and more robust. For this, introduced a new `skipping`
parameter set to Multiverse, which is evaluated as part of the universe
setup functions and granularly allows to continue, skip or raise
depending on whether the universe output folder, config file, or
universe output files exist.