Write report files for run-existing as well
This required a little overhaul of the internal reporting
implementation, removing explicit mentions of "joined" runs and instead
focussing on whether the reporter is attached to a
DistributedMultiverse.

Also extends the Multiverse and reporter with descriptive "run tags",
which can help understand what kind of run is being carried out.

Also fixes a bug in run directory detection.